Add files via upload
Last semester, I had the opportunity to work on a project that involved creating an infographic. As someone who was new to the field of artificial intelligence (AI), I decided to challenge myself by incorporating AI into the project.

Using AI tools, I was able to collect and analyze large amounts of data to create visual representations that effectively communicated the key insights and takeaways of the project. By leveraging AI algorithms, I was able to process complex data sets quickly and efficiently, which helped me save time and focus more on the creative aspects of the project.

Overall, the experience of using AI to create an infographic was both challenging and rewarding. It not only allowed me to expand my skill set and knowledge of AI, but also helped me create a more impactful and engaging project that effectively conveyed the message I wanted to communicate.