Add files via upload
During the last semester, I had the opportunity to create a brochure for a project I was working on. It was a challenging task, but one that I was excited to take on because it was my first time using Adobe InDesign (ID) for a project.

As someone who was new to the software, I had to spend some time getting familiar with the different features and tools that it offered. But once I got the hang of it, I was able to use ID to create a brochure that effectively communicated the project's message and goals.

Using ID, I was able to experiment with different layouts, fonts, and colors to create a design that was both visually appealing and easy to read. I also appreciated the software's ability to handle large amounts of text and images, which made it easier to organize and structure the content of the brochure.

Overall, I found the experience of using ID to create a brochure to be very rewarding. Not only did I learn a new software tool, but I also produced a high-quality brochure that effectively communicated the message of the project. I look forward to continuing to use ID in my future design work.